Refactor script codegen
