+++ packages/language-core/lib/codegen/script/componentSetupReturns.ts
@@ -1,0 +1,16 @@
+import type { ScriptSetupRanges } from '../../parsers/scriptSetupRanges';
+import type { Code } from '../../types';
+import { newLine } from '../common';
+
+export function* generateComponentSetupReturns(scriptSetupRanges: ScriptSetupRanges): Generator<Code> {
+	// fill $props
+	if (scriptSetupRanges.props.define) {
+		// NOTE: defineProps is inaccurate for $props
+		yield `$props: __VLS_makeOptional(${scriptSetupRanges.props.name ?? `__VLS_props`}),${newLine}`;
+		yield `...${scriptSetupRanges.props.name ?? `__VLS_props`},${newLine}`;
+	}
+	// fill $emit
+	if (scriptSetupRanges.emits.define) {
+		yield `$emit: ${scriptSetupRanges.emits.name ?? '__VLS_emit'},${newLine}`;
+	}
+}
