+++ packages/language-core/lib/codegen/template/index.ts
@@ -1,56 +1,16 @@
 import * as CompilerDOM from '@vue/compiler-dom';
 import { camelize, capitalize } from '@vue/shared';
 import type * as ts from 'typescript';
-import type { Code, Sfc, VueCodeInformation, VueCompilerOptions } from "../../types";
+import type { Code, Sfc, VueCodeInformation, VueCompilerOptions } from '../../types';
 import { hyphenateTag } from '../../utils/shared';
-import { endOfLine, newLine, variableNameRegex, wrapWith } from "../common";
+import { endOfLine, newLine, variableNameRegex, wrapWith } from '../common';
 import { generateCamelized } from './camelized';
+import { createTemplateCodegenContext } from './context';
 import { getCanonicalComponentName, getPossibleOriginalComponentNames } from './element';
 import { generateObjectProperty } from './objectProperty';
 import { generatePropertyAccess } from './propertyAccess';
 import { generateStringLiteralKey } from './stringLiteralKey';
-import { generateTemplateNode, getVForNode } from './templateNode';
-
-const _codeFeatures = {
-	all: {
-		verification: true,
-		completion: true,
-		semantic: true,
-		navigation: true,
-	} as VueCodeInformation,
-	verification: {
-		verification: true,
-	} as VueCodeInformation,
-	completion: {
-		completion: true,
-	} as VueCodeInformation,
-	additionalCompletion: {
-		completion: { isAdditional: true },
-	} as VueCodeInformation,
-	navigation: {
-		navigation: true,
-	} as VueCodeInformation,
-	navigationAndCompletion: {
-		navigation: true,
-	} as VueCodeInformation,
-	withoutHighlight: {
-		semantic: { shouldHighlight: () => false },
-		verification: true,
-		navigation: true,
-		completion: true,
-	} as VueCodeInformation,
-	withoutHighlightAndCompletion: {
-		semantic: { shouldHighlight: () => false },
-		verification: true,
-		navigation: true,
-	} as VueCodeInformation,
-	withoutHighlightAndCompletionAndNavigation: {
-		semantic: { shouldHighlight: () => false },
-		verification: true,
-	} as VueCodeInformation,
-};
-
-export type TemplateCodegenContext = ReturnType<typeof createTemplateCodegenContext>;
+import { generateTemplateChild, getVForNode } from './templateChild';
 
 export interface TemplateCodegenOptions {
 	ts: typeof ts;
@@ -64,7 +24,7 @@ export interface TemplateCodegenOptions {
 	propsAssignName?: string;
 }
 
-export function* generate(options: TemplateCodegenOptions) {
+export function* generateTemplate(options: TemplateCodegenOptions) {
 	const ctx = createTemplateCodegenContext();
 	const { componentTagNameOffsets, elementTagNameOffsets } = collectTagOffsets();
 
@@ -80,7 +40,7 @@ export function* generate(options: TemplateCodegenOptions) {
 	yield* generatePreResolveComponents();
 
 	if (options.template.ast) {
-		yield* generateTemplateNode(options, ctx, options.template.ast, undefined, undefined, undefined);
+		yield* generateTemplateChild(options, ctx, options.template.ast, undefined, undefined, undefined);
 	}
 
 	yield* generateStyleScopedClasses();
@@ -316,152 +276,6 @@ function camelizeComponentName(newName: string) {
 
 function getTagRenameApply(oldName: string) {
 	return oldName === hyphenateTag(oldName) ? hyphenateTag : undefined;
-}
-
-export function createTemplateCodegenContext() {
-	let ignoredError = false;
-	let expectErrorToken: {
-		errors: number;
-		node: CompilerDOM.CommentNode;
-	} | undefined;
-	let variableId = 0;
-
-	const codeFeatures = new Proxy(_codeFeatures, {
-		get(target, key: keyof typeof _codeFeatures) {
-			const data = target[key];
-			if (data.verification) {
-				if (ignoredError) {
-					return {
-						...data,
-						verification: false,
-					};
-				}
-				if (expectErrorToken) {
-					const token = expectErrorToken;
-					if (typeof data.verification !== 'object' || !data.verification.shouldReport) {
-						return {
-							...data,
-							verification: {
-								shouldReport: () => {
-									token.errors++;
-									return false;
-								},
-							},
-						};
-					}
-				}
-			}
-			return data;
-		},
-	});
-	const localVars = new Map<string, number>();
-	const accessGlobalVariables = new Map<string, Set<number>>();
-	const slots: {
-		name: string;
-		loc?: number;
-		tagRange: [number, number];
-		varName: string;
-		nodeLoc: any;
-	}[] = [];
-	const dynamicSlots: {
-		expVar: string;
-		varName: string;
-	}[] = [];
-	const hasSlotElements = new Set<CompilerDOM.ElementNode>();;
-	const blockConditions: string[] = [];
-	const usedComponentCtxVars = new Set<string>();
-	const scopedClasses: { className: string, offset: number; }[] = [];
-
-	return {
-		slots,
-		dynamicSlots,
-		codeFeatures,
-		accessGlobalVariables,
-		hasSlotElements,
-		blockConditions,
-		usedComponentCtxVars,
-		scopedClasses,
-		accessGlobalVariable(name: string, offset?: number) {
-			let arr = accessGlobalVariables.get(name);
-			if (!arr) {
-				accessGlobalVariables.set(name, arr = new Set());
-			}
-			if (offset !== undefined) {
-				arr.add(offset);
-			}
-		},
-		hasLocalVariable: (name: string) => {
-			return !!localVars.get(name);
-		},
-		addLocalVariable: (name: string) => {
-			localVars.set(name, (localVars.get(name) ?? 0) + 1);
-		},
-		removeLocalVariable: (name: string) => {
-			localVars.set(name, localVars.get(name)! - 1);
-		},
-		getInternalVariable: () => {
-			return `__VLS_${variableId++}`;
-		},
-		ignoreError: function* (): Generator<Code> {
-			if (!ignoredError) {
-				ignoredError = true;
-				yield `// @vue-ignore start${newLine}`;
-			}
-		},
-		expectError: function* (prevNode: CompilerDOM.CommentNode): Generator<Code> {
-			if (!expectErrorToken) {
-				expectErrorToken = {
-					errors: 0,
-					node: prevNode,
-				};
-				yield `// @vue-expect-error start${newLine}`;
-			}
-		},
-		resetDirectiveComments: function* (endStr: string): Generator<Code> {
-			if (expectErrorToken) {
-				const token = expectErrorToken;
-				yield* wrapWith(
-					expectErrorToken.node.loc.start.offset,
-					expectErrorToken.node.loc.end.offset,
-					{
-						verification: {
-							shouldReport: () => token.errors === 0,
-						},
-					},
-					`// @ts-expect-error __VLS_TS_EXPECT_ERROR`,
-				);
-				yield `${newLine}${endOfLine}`;
-				expectErrorToken = undefined;
-				yield `// @vue-expect-error ${endStr}${newLine}`;
-			}
-			if (ignoredError) {
-				ignoredError = false;
-				yield `// @vue-ignore ${endStr}${newLine}`;
-			}
-		},
-		generateAutoImportCompletion: function* (): Generator<Code> {
-			const all = [...accessGlobalVariables.entries()];
-			if (!all.some(([_, offsets]) => offsets.size)) {
-				yield `// no auto imports${endOfLine}`;
-				return;
-			}
-			yield `// @ts-ignore${newLine}`; // #2304
-			yield `[`;
-			for (const [varName, offsets] of all) {
-				for (const offset of offsets) {
-					yield [
-						varName,
-						'template',
-						offset,
-						codeFeatures.additionalCompletion,
-					];
-					yield `,`;
-				}
-				offsets.clear();
-			}
-			yield `]${endOfLine}`;
-		}
-	};
 }
 
 export function isFragment(node: CompilerDOM.IfNode | CompilerDOM.ForNode) {
