+++ packages/language-core/lib/codegen/template/interpolation.ts
@@ -3,7 +3,8 @@ import type * as ts from 'typescript';
 import { getNodeText, getStartEnd } from '../../parsers/scriptSetupRanges';
 import type { Code, VueCodeInformation, VueCompilerOptions } from '../../types';
 import { collectVars, createTsAst } from '../common';
-import type { TemplateCodegenContext, TemplateCodegenOptions } from './index';
+import type { TemplateCodegenContext } from './context';
+import type { TemplateCodegenOptions } from './index';
 
 export function* generateInterpolation(
 	options: TemplateCodegenOptions,
