+++ packages/language-core/lib/codegen/template/stringLiteralKey.ts
@@ -1,5 +1,5 @@
-import type { Code, VueCodeInformation } from "../../types";
-import { combineLastMapping, wrapWith } from "../common";
+import type { Code, VueCodeInformation } from '../../types';
+import { combineLastMapping, wrapWith } from '../common';
 
 export function* generateStringLiteralKey(code: string, offset?: number, info?: VueCodeInformation): Generator<Code> {
 	if (offset === undefined || !info) {
