+++ packages/language-service/tests/format/1210.spec.ts
@@ -5,7 +5,7 @@ defineFormatTest({
 	languageId: 'vue',
 	input: `
 <script setup lang="ts">
-import { ref } from "vue";
+import { ref } from 'vue';
 
 const count = ref(0);
 </script>
