+++ test-workspace/component-meta/options-api/component.ts
@@ -1,4 +1,4 @@
-import { defineComponent } from "vue";
+import { defineComponent } from 'vue';
 
 interface SubmitPayload {
 	/**
