+++ test-workspace/component-meta/ts-component/component.ts
@@ -1,5 +1,5 @@
-import { h, defineComponent } from "vue";
-import { MyProps } from "./PropDefinitions";
+import { h, defineComponent } from 'vue';
+import { MyProps } from './PropDefinitions';
 
 export default defineComponent((props: MyProps) => {
 	return () => h('pre', JSON.stringify(props, null, 2));
