+++ test-workspace/tsc/vue3/#2399/main.vue
@@ -3,13 +3,13 @@
 </template>
 
 <script setup lang="ts">
-import { ref } from "vue";
+import { ref } from 'vue';
 withDefaults(
 	defineProps<{
-		type?: "checkbox" | "radio";
+		type?: 'checkbox' | 'radio';
 	}>(),
 	{
-		type: "checkbox",
+		type: 'checkbox',
 	}
 );
 const defaultSelection: string[] = [];
